+++ src/main/java/org/synyx/urlaubsverwaltung/availability/api/AvailabilityApiController.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,15 +16,15 @@ import org.synyx.urlaubsverwaltung.person.PersonService;
 import org.synyx.urlaubsverwaltung.security.SecurityRules;
 
 import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
 import java.util.Optional;
 
 
 @Api("Availabilities: Get all availabilities for a certain period")
 @RestController("restApiAvailabilityController")
-@RequestMapping("/api")
+@RequestMapping("/api/")
 public class AvailabilityApiController {
 
+    public static final String AVAILABILITIES = "availabilities";
     private final PersonService personService;
     private final AvailabilityService availabilityService;
 
@@ -39,36 +40,29 @@ public class AvailabilityApiController {
         notes =
             "Get all availabilities for a certain period and person. Maximum allowed period per request is one month."
     )
-    @GetMapping("/availabilities")
+    @GetMapping(AVAILABILITIES)
     @PreAuthorize(SecurityRules.IS_OFFICE)
-    public AvailabilityList personsAvailabilities(
+    public AvailabilityListDto personsAvailabilities(
         @ApiParam(value = "start of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_FIRST_DAY_OF_YEAR)
         @RequestParam("from")
-            String startDateString,
-        @ApiParam(value = "end of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_LAST_DAY_OF_YEAR)
+        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+            LocalDate startDate,
+        @ApiParam(value = "end of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_LAST_DAY_OF_MONTH)
         @RequestParam("to")
-            String endDateString,
-        @ApiParam(value = "login name of the person")
+        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+            LocalDate endDate,
+        @ApiParam(value = "id of the person")
         @RequestParam(value = "person")
-            String personLoginName) {
+            Integer personId) {
 
-        final Optional<Person> optionalPerson = personService.getPersonByLogin(personLoginName);
+        final Optional<Person> optionalPerson = personService.getPersonByID(personId);
 
-        if (!optionalPerson.isPresent()) {
-            throw new IllegalArgumentException("No person found for loginName = " + personLoginName);
-        }
-
-        final LocalDate startDate;
-        final LocalDate endDate;
-        try {
-            startDate = LocalDate.parse(startDateString);
-            endDate = LocalDate.parse(endDateString);
-        } catch (DateTimeParseException exception) {
-            throw new IllegalArgumentException(exception.getMessage());
+        if (optionalPerson.isEmpty()) {
+            throw new IllegalArgumentException("No person found for id = " + personId);
         }
 
         if (startDate.isAfter(endDate)) {
-            throw new IllegalArgumentException("startdate " + startDateString + " must not be after endDate " + endDateString);
+            throw new IllegalArgumentException("startdate " + startDate + " must not be after endDate " + endDate);
         }
 
         boolean requestedDateRangeIsMoreThanOneMonth = startDate.minusDays(1).isBefore(endDate.minusMonths(1));