+++ src/main/java/org/synyx/urlaubsverwaltung/person/api/PersonApiController.java
@@ -2,25 +2,26 @@ package org.synyx.urlaubsverwaltung.person.api;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import org.apache.http.client.utils.URIBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.hateoas.Link;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.synyx.urlaubsverwaltung.availability.api.AvailabilityApiController;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
 import org.synyx.urlaubsverwaltung.security.SecurityRules;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
+import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
+import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
+import static org.synyx.urlaubsverwaltung.availability.api.AvailabilityApiController.AVAILABILITIES;
 
 @Api("Persons: Get information about the persons of the application")
 @RestController("restApiPersonController")
@@ -43,10 +44,10 @@ public class PersonApiController {
     )
     @GetMapping
     @PreAuthorize(SecurityRules.IS_OFFICE)
-    public ResponseEntity<List<PersonResponse>> persons(HttpServletRequest request) {
+    public ResponseEntity<List<PersonResponse>> persons() {
 
         List<PersonResponse> persons = personService.getActivePersons().stream()
-            .map(p -> createPersonResponse(p, request))
+            .map(this::createPersonResponse)
             .collect(toList());
 
         return new ResponseEntity<>(persons, OK);
@@ -55,22 +56,18 @@ public class PersonApiController {
     @ApiOperation(value = "Get one active person by id", notes = "Get one active person by id")
     @GetMapping(PERSON_URL)
     @PreAuthorize(SecurityRules.IS_OFFICE)
-    public ResponseEntity<PersonResponse> getPerson(@PathVariable Integer id, HttpServletRequest request) {
+    public ResponseEntity<PersonResponse> getPerson(@PathVariable Integer id) {
 
         return personService.getPersonByID(id)
-            .map(value -> new ResponseEntity<>(createPersonResponse(value, request), OK))
+            .map(value -> new ResponseEntity<>(createPersonResponse(value), OK))
             .orElseGet(() -> new ResponseEntity<>(NOT_FOUND));
     }
 
-    private PersonResponse createPersonResponse(Person person, HttpServletRequest request) {
+    private PersonResponse createPersonResponse(Person person) {
         final PersonResponse personResponse = PersonResponseMapper.mapToResponse(person);
-        final String baseUri = new URIBuilder()
-            .setScheme(request.getScheme())
-            .setPort(request.getServerPort())
-            .setHost(request.getServerName())
-            .setPath(ROOT_URL)
-            .toString();
-        personResponse.add(new Link(baseUri + "/" + person.getId()).withSelfRel());
+        personResponse.add(linkTo(methodOn(PersonApiController.class).getPerson(person.getId())).withSelfRel());
+        personResponse.add(linkTo(methodOn(AvailabilityApiController.class).personsAvailabilities(null, null, person.getId()))
+                .withRel(AVAILABILITIES));
         return personResponse;
     }
 }