+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade spring boot parent to 2.1.7
 * Upgrade NodeJS to 12
 * Upgrade date-fns from 1.x to 2.0.1
+* API-Change: `api/availability` now uses `personId` instead of `loginName`. The id of a person can be discovered via `api/person` [736](https://github.com/synyx/urlaubsverwaltung/issues/736)
 
 ### [urlaubsverwaltung-3.0.0.RC4](https://github.com/synyx/urlaubsverwaltung/releases/tag/urlaubsverwaltung-3.0.0.RC4)
 * Fix login parameter security